Fixed type

+++ MusicService/DTOs/TrackMachineLearningDTO.cs
@@ -2,16 +2,16 @@ namespace MusicService.DTOs
 {
     public class TrackMachineLearningDTO
     {
-        public float? Popularity { get; set; }
-        public float? Acousticness { get; set; }
-        public float? Danceability { get; set; }
-        public float? DurationMs { get; set; }
-        public float? Tempo { get; set; }
-        public float? Energy { get; set; }
-        public float? Instrumentalness { get; set;}
-        public float? Liveness { get; set; }
-        public float? Loudness { get; set; }
-        public float? Speechiness { get; set; }
-        public float? Valence { get; set; }
+        public double? Popularity { get; set; }
+        public double? Acousticness { get; set; }
+        public double? Danceability { get; set; }
+        public double? DurationMs { get; set; }
+        public double? Tempo { get; set; }
+        public double? Energy { get; set; }
+        public double? Instrumentalness { get; set;}
+        public double? Liveness { get; set; }
+        public double? Loudness { get; set; }
+        public double? Speechiness { get; set; }
+        public double? Valence { get; set; }
     }
 }
